WIP: shuffle everything around
