+++ server/graphql-engine.cabal
@@ -300,12 +300,25 @@ library
                      , Data.Time.Clock.Units
                      , Data.URL.Template
                      , Hasura.App
+
+                     , Hasura.Backends.Postgres.Connection
+                     , Hasura.Backends.Postgres.Execute.Query
+                     , Hasura.Backends.Postgres.Execute.Prepare
+                     , Hasura.Backends.Postgres.Execute.Mutation
+                     , Hasura.Backends.Postgres.Execute.RemoteJoin
+                     , Hasura.Backends.Postgres.Translate.Delete
+                     , Hasura.Backends.Postgres.Translate.Insert
+                     , Hasura.Backends.Postgres.Translate.Mutation
+                     , Hasura.Backends.Postgres.Translate.RemoteJoin
+                     , Hasura.Backends.Postgres.Translate.Returning
+                     , Hasura.Backends.Postgres.Translate.Select
+                     , Hasura.Backends.Postgres.Translate.Update
                      , Hasura.Backends.Postgres.SQL.DML
                      , Hasura.Backends.Postgres.SQL.Error
-                     , Hasura.Backends.Postgres.Connection
                      , Hasura.Backends.Postgres.SQL.Rewrite
                      , Hasura.Backends.Postgres.SQL.Types
                      , Hasura.Backends.Postgres.SQL.Value
+
                      -- Exposed for benchmark:
                      , Hasura.Cache.Bounded
                      , Hasura.Logging
@@ -412,20 +425,17 @@ library
                      , Hasura.RQL.DDL.Utils
                      , Hasura.RQL.DDL.EventTrigger
                      , Hasura.RQL.DDL.ScheduledTrigger
+                     , Hasura.RQL.DML.Count
                      , Hasura.RQL.DML.Delete
                      , Hasura.RQL.DML.Delete.Types
-                     , Hasura.RQL.DML.Internal
                      , Hasura.RQL.DML.Insert
                      , Hasura.RQL.DML.Insert.Types
-                     , Hasura.RQL.DML.Mutation
-                     , Hasura.RQL.DML.RemoteJoin
-                     , Hasura.RQL.DML.Returning
+                     , Hasura.RQL.DML.Internal
+                     , Hasura.RQL.DML.RemoteJoin.Types
                      , Hasura.RQL.DML.Returning.Types
-                     , Hasura.RQL.DML.Select.Internal
                      , Hasura.RQL.DML.Select.Types
                      , Hasura.RQL.DML.Update
                      , Hasura.RQL.DML.Update.Types
-                     , Hasura.RQL.DML.Count
                      , Hasura.RQL.GBoolExp
                      , Hasura.GraphQL.Explain
                      , Hasura.GraphQL.Execute.Action
