+++ server/src-lib/Hasura/Backends/Postgres/Translate/Mutation.hs
@@ -1,0 +1,72 @@
+module Hasura.Backends.Postgres.Translate.Mutation
+  ( mkSelCTEFromColVals
+  )
+where
+
+import           Hasura.Prelude
+
+import qualified Data.Environment                    as Env
+import qualified Data.HashMap.Strict                 as Map
+import qualified Data.Sequence                       as DS
+import qualified Database.PG.Query                   as Q
+import qualified Network.HTTP.Client                 as HTTP
+import qualified Network.HTTP.Types                  as N
+
+import           Data.Text.Extended
+
+import qualified Hasura.Backends.Postgres.SQL.DML    as S
+import qualified Hasura.Tracing                      as Tracing
+
+import           Hasura.Backends.Postgres.Connection
+import           Hasura.Backends.Postgres.SQL.Types
+import           Hasura.Backends.Postgres.SQL.Value
+import           Hasura.EncJSON
+import           Hasura.RQL.DML.Internal
+import           Hasura.RQL.DML.RemoteJoin.Types
+import           Hasura.RQL.DML.Returning.Types
+import           Hasura.RQL.DML.Select
+import           Hasura.RQL.Instances                ()
+import           Hasura.RQL.Types
+import           Hasura.SQL.Types
+import           Hasura.Server.Version               (HasVersion)
+import           Hasura.Session
+
+
+-- | Note:- Using sorted columns is necessary to enable casting the rows returned by VALUES expression to table type.
+-- For example, let's consider the table, `CREATE TABLE test (id serial primary key, name text not null, age int)`.
+-- The generated values expression should be in order of columns;
+-- `SELECT ("row"::table).* VALUES (1, 'Robert', 23) AS "row"`.
+mkSelCTEFromColVals
+  :: (MonadError QErr m)
+  => QualifiedTable -> [ColumnInfo 'Postgres] -> [ColumnValues TxtEncodedPGVal] -> m S.CTE
+mkSelCTEFromColVals qt allCols colVals =
+  S.CTESelect <$> case colVals of
+    [] -> return selNoRows
+    _  -> do
+      tuples <- mapM mkTupsFromColVal colVals
+      let fromItem = S.FIValues (S.ValuesExp tuples) (S.Alias rowAlias) Nothing
+      return S.mkSelect
+        { S.selExtr = [extractor]
+        , S.selFrom = Just $ S.FromExp [fromItem]
+        }
+  where
+    rowAlias = Identifier "row"
+    extractor = S.selectStar' $ S.QualifiedIdentifier rowAlias $ Just $ S.TypeAnn $ toSQLTxt qt
+    sortedCols = sortCols allCols
+    mkTupsFromColVal colVal =
+      fmap S.TupleExp $ forM sortedCols $ \ci -> do
+        let pgCol = pgiColumn ci
+        val <- onNothing (Map.lookup pgCol colVal) $
+          throw500 $ "column " <> pgCol <<> " not found in returning values"
+        pure $ txtEncodedToSQLExp (pgiType ci) val
+
+    selNoRows =
+      S.mkSelect { S.selExtr = [S.selectStar]
+                 , S.selFrom = Just $ S.mkSimpleFromExp qt
+                 , S.selWhere = Just $ S.WhereFrag $ S.BELit False
+                 }
+
+    txtEncodedToSQLExp colTy = \case
+      TENull          -> S.SENull
+      TELit textValue ->
+        S.withTyAnn (unsafePGColumnToRepresentation colTy) $ S.SELit textValue
