+++ server/src-lib/Hasura/RQL/DML/Delete/Types.hs
@@ -1,5 +1,7 @@
 module Hasura.RQL.DML.Delete.Types where
 
+import           Hasura.Prelude
+
 import qualified Hasura.Backends.Postgres.SQL.DML   as S
 
 import           Hasura.Backends.Postgres.SQL.Types
@@ -7,6 +9,7 @@ import           Hasura.RQL.DML.Returning.Types
 import           Hasura.RQL.Types.BoolExp
 import           Hasura.RQL.Types.Column
 import           Hasura.SQL.Backend
+
 
 data AnnDelG (b :: Backend) v
   = AnnDel
@@ -17,3 +20,16 @@ data AnnDelG (b :: Backend) v
   }
 
 type AnnDel b = AnnDelG b S.SQLExp
+
+traverseAnnDel
+  :: (Applicative f)
+  => (a -> f b)
+  -> AnnDelG backend a
+  -> f (AnnDelG backend b)
+traverseAnnDel f annUpd =
+  AnnDel tn
+  <$> ((,) <$> traverseAnnBoolExp f whr <*> traverseAnnBoolExp f fltr)
+  <*> traverseMutationOutput f mutOutput
+  <*> pure allCols
+  where
+    AnnDel tn (whr, fltr) mutOutput allCols = annUpd
