+++ server/src-lib/Hasura/RQL/DML/Select.hs
@@ -1,32 +1,28 @@
 module Hasura.RQL.DML.Select
   ( selectP2
   , convSelectQuery
-  , asSingleRowJsonResp
   , runSelect
-  , selectQuerySQL
-  , selectAggregateQuerySQL
-  , connectionSelectQuerySQL
-  , module Hasura.RQL.DML.Select.Internal
   )
 where
 
 import           Hasura.Prelude
 
-import qualified Data.HashSet                       as HS
-import qualified Data.List.NonEmpty                 as NE
-import qualified Data.Sequence                      as DS
-import qualified Database.PG.Query                  as Q
+import qualified Data.HashSet                              as HS
+import qualified Data.List.NonEmpty                        as NE
+import qualified Data.Sequence                             as DS
+import qualified Database.PG.Query                         as Q
 
 import           Data.Aeson.Types
 import           Data.Text.Extended
-import           Instances.TH.Lift                  ()
+import           Instances.TH.Lift                         ()
 
-import qualified Hasura.Backends.Postgres.SQL.DML   as S
+import qualified Hasura.Backends.Postgres.SQL.DML          as S
 
 import           Hasura.Backends.Postgres.SQL.Types
+import           Hasura.Backends.Postgres.Translate.Select
 import           Hasura.EncJSON
 import           Hasura.RQL.DML.Internal
-import           Hasura.RQL.DML.Select.Internal
+import           Hasura.RQL.DML.Select.Types
 import           Hasura.RQL.Types
 import           Hasura.SQL.Types
 
@@ -283,23 +279,6 @@ selectP2 jsonAggSelect (sel, p) =
   <$> Q.rawQE dmlTxErrorHandler (Q.fromBuilder selectSQL) (toList p) True
   where
     selectSQL = toSQL $ mkSQLSelect jsonAggSelect sel
-
-selectQuerySQL :: JsonAggSelect -> AnnSimpleSel 'Postgres -> Q.Query
-selectQuerySQL jsonAggSelect sel =
-  Q.fromBuilder $ toSQL $ mkSQLSelect jsonAggSelect sel
-
-selectAggregateQuerySQL :: AnnAggregateSelect 'Postgres -> Q.Query
-selectAggregateQuerySQL =
-  Q.fromBuilder . toSQL . mkAggregateSelect
-
-connectionSelectQuerySQL :: ConnectionSelect 'Postgres S.SQLExp -> Q.Query
-connectionSelectQuerySQL =
-  Q.fromBuilder . toSQL . mkConnectionSelect
-
-asSingleRowJsonResp :: Q.Query -> [Q.PrepArg] -> Q.TxE QErr EncJSON
-asSingleRowJsonResp query args =
-  encJFromBS . runIdentity . Q.getRow
-  <$> Q.rawQE dmlTxErrorHandler query args True
 
 phaseOne
   :: (QErrM m, UserInfoM m, CacheRM m, HasSQLGenCtx m)
