+++ server/src-lib/Hasura/RQL/DML/Update.hs
@@ -1,104 +1,42 @@
 module Hasura.RQL.DML.Update
-  ( AnnUpdG(..)
-  , traverseAnnUpd
-  , execUpdateQuery
-  , updateOperatorText
-  , runUpdate
+  ( runUpdate
   ) where
 
 import           Hasura.Prelude
 
-import qualified Data.Environment                   as Env
-import qualified Data.HashMap.Strict                as M
-import qualified Data.Sequence                      as DS
-import qualified Database.PG.Query                  as Q
+import qualified Data.Environment                              as Env
+import qualified Data.HashMap.Strict                           as M
+import qualified Data.Sequence                                 as DS
+import qualified Database.PG.Query                             as Q
 
 import           Data.Aeson.Types
 import           Data.Text.Extended
-import           Instances.TH.Lift                  ()
+import           Instances.TH.Lift                             ()
 
-import qualified Hasura.Backends.Postgres.SQL.DML   as S
-import qualified Hasura.Tracing                     as Tracing
+import qualified Hasura.Backends.Postgres.SQL.DML              as S
+import qualified Hasura.Tracing                                as Tracing
 
+import           Hasura.Backends.Postgres.Connection
+import           Hasura.Backends.Postgres.Execute.Mutation
+import           Hasura.Backends.Postgres.Execute.Prepare
+import           Hasura.Backends.Postgres.Execute.Query
 import           Hasura.Backends.Postgres.SQL.Types
+import           Hasura.Backends.Postgres.Translate.Delete
+import           Hasura.Backends.Postgres.Translate.Insert
+import           Hasura.Backends.Postgres.Translate.Mutation
+import           Hasura.Backends.Postgres.Translate.RemoteJoin
+import           Hasura.Backends.Postgres.Translate.Returning
+import           Hasura.Backends.Postgres.Translate.Select
+import           Hasura.Backends.Postgres.Translate.Update
 import           Hasura.EncJSON
-import           Hasura.RQL.DML.Insert              (insertCheckExpr)
 import           Hasura.RQL.DML.Internal
-import           Hasura.RQL.DML.Mutation
-import           Hasura.RQL.DML.Returning
 import           Hasura.RQL.DML.Update.Types
 import           Hasura.RQL.GBoolExp
-import           Hasura.RQL.Instances               ()
+import           Hasura.RQL.Instances                          ()
 import           Hasura.RQL.Types
-import           Hasura.Server.Version              (HasVersion)
+import           Hasura.Server.Version                         (HasVersion)
 import           Hasura.Session
 
-
--- NOTE: This function can be improved, because we use
--- the literal values defined below in the 'updateOperators'
--- function in 'Hasura.GraphQL.Schema.Mutation'. It would
--- be nice if we could avoid duplicating the string literal
--- values
-updateOperatorText :: UpdOpExpG a -> Text
-updateOperatorText (UpdSet          _) = "_set"
-updateOperatorText (UpdInc          _) = "_inc"
-updateOperatorText (UpdAppend       _) = "_append"
-updateOperatorText (UpdPrepend      _) = "_prepend"
-updateOperatorText (UpdDeleteKey    _) = "_delete_key"
-updateOperatorText (UpdDeleteElem   _) = "_delete_elem"
-updateOperatorText (UpdDeleteAtPath _) = "_delete_at_path"
-
-traverseAnnUpd
-  :: (Applicative f)
-  => (a -> f b)
-  -> AnnUpdG backend a
-  -> f (AnnUpdG backend b)
-traverseAnnUpd f annUpd =
-  AnnUpd tn
-  <$> traverse (traverse $ traverse f) opExps
-  <*> ((,) <$> traverseAnnBoolExp f whr <*> traverseAnnBoolExp f fltr)
-  <*> traverseAnnBoolExp f chk
-  <*> traverseMutationOutput f mutOutput
-  <*> pure allCols
-  where
-    AnnUpd tn opExps (whr, fltr) chk mutOutput allCols = annUpd
-
-mkUpdateCTE
-  :: AnnUpd 'Postgres -> S.CTE
-mkUpdateCTE (AnnUpd tn opExps (permFltr, wc) chk _ columnsInfo) =
-  S.CTEUpdate update
-  where
-    update =
-      S.SQLUpdate tn setExp Nothing tableFltr
-        . Just
-        . S.RetExp
-        $ [ S.selectStar
-          , S.Extractor (insertCheckExpr "update check constraint failed" checkExpr) Nothing
-          ]
-    setExp    = S.SetExp $ map (expandOperator columnsInfo) opExps
-    tableFltr = Just $ S.WhereFrag tableFltrExpr
-    tableFltrExpr = toSQLBoolExp (S.QualTable tn) $ andAnnBoolExps permFltr wc
-    checkExpr = toSQLBoolExp (S.QualTable tn) chk
-
-expandOperator :: [ColumnInfo 'Postgres] -> (PGCol, UpdOpExpG S.SQLExp) -> S.SetExpItem
-expandOperator infos (column, op) = S.SetExpItem $ (column,) $ case op of
-  UpdSet          e -> e
-  UpdInc          e -> S.mkSQLOpExp S.incOp               identifier (asNum  e)
-  UpdAppend       e -> S.mkSQLOpExp S.jsonbConcatOp       identifier (asJSON e)
-  UpdPrepend      e -> S.mkSQLOpExp S.jsonbConcatOp       (asJSON e) identifier
-  UpdDeleteKey    e -> S.mkSQLOpExp S.jsonbDeleteOp       identifier (asText e)
-  UpdDeleteElem   e -> S.mkSQLOpExp S.jsonbDeleteOp       identifier (asInt  e)
-  UpdDeleteAtPath a -> S.mkSQLOpExp S.jsonbDeleteAtPathOp identifier (asArray a)
-  where
-    identifier = S.SEIdentifier $ toIdentifier column
-    asInt  e   = S.SETyAnn e S.intTypeAnn
-    asText e   = S.SETyAnn e S.textTypeAnn
-    asJSON e   = S.SETyAnn e S.jsonbTypeAnn
-    asArray a  = S.SETyAnn (S.SEArray a) S.textArrTypeAnn
-    asNum  e   = S.SETyAnn e $
-      case find (\info -> pgiColumn info == column) infos <&> pgiType of
-        Just (PGColumnScalar s) -> S.mkTypeAnn $ PGTypeScalar s
-        _                       -> S.numericTypeAnn
 
 convInc
   :: (QErrM m)
@@ -249,24 +187,6 @@ validateUpdateQuery
   => UpdateQuery -> m (AnnUpd 'Postgres, DS.Seq Q.PrepArg)
 validateUpdateQuery =
   runDMLP1T . validateUpdateQueryWith sessVarFromCurrentSetting binRHSBuilder
-
-execUpdateQuery
-  ::
-  ( HasVersion
-  , MonadTx m
-  , MonadIO m
-  , Tracing.MonadTrace m
-  )
-  => Env.Environment
-  -> Bool
-  -> Maybe MutationRemoteJoinCtx
-  -> (AnnUpd 'Postgres, DS.Seq Q.PrepArg)
-  -> m EncJSON
-execUpdateQuery env strfyNum remoteJoinCtx (u, p) =
-  runMutation env $ mkMutation remoteJoinCtx (uqp1Table u) (updateCTE, p)
-                (uqp1Output u) (uqp1AllCols u) strfyNum
-  where
-    updateCTE = mkUpdateCTE u
 
 runUpdate
   :: ( HasVersion, QErrM m, UserInfoM m, CacheRM m
