+++ server/src-lib/Hasura/RQL/DML/Update/Types.hs
@@ -36,3 +36,33 @@ data UpdOpExpG v = UpdSet !v
                  | UpdDeleteElem !v
                  | UpdDeleteAtPath ![v]
                  deriving (Functor, Foldable, Traversable, Generic, Data)
+
+
+-- NOTE: This function can be improved, because we use
+-- the literal values defined below in the 'updateOperators'
+-- function in 'Hasura.GraphQL.Schema.Mutation'. It would
+-- be nice if we could avoid duplicating the string literal
+-- values
+updateOperatorText :: UpdOpExpG a -> Text
+updateOperatorText (UpdSet          _) = "_set"
+updateOperatorText (UpdInc          _) = "_inc"
+updateOperatorText (UpdAppend       _) = "_append"
+updateOperatorText (UpdPrepend      _) = "_prepend"
+updateOperatorText (UpdDeleteKey    _) = "_delete_key"
+updateOperatorText (UpdDeleteElem   _) = "_delete_elem"
+updateOperatorText (UpdDeleteAtPath _) = "_delete_at_path"
+
+traverseAnnUpd
+  :: (Applicative f)
+  => (a -> f b)
+  -> AnnUpdG backend a
+  -> f (AnnUpdG backend b)
+traverseAnnUpd f annUpd =
+  AnnUpd tn
+  <$> traverse (traverse $ traverse f) opExps
+  <*> ((,) <$> traverseAnnBoolExp f whr <*> traverseAnnBoolExp f fltr)
+  <*> traverseAnnBoolExp f chk
+  <*> traverseMutationOutput f mutOutput
+  <*> pure allCols
+  where
+    AnnUpd tn opExps (whr, fltr) chk mutOutput allCols = annUpd
